connect TelegramBot to DB with create sequence

+++ src/main/java/ru/complexhex/scrapernews/entity/User.java
@@ -13,11 +13,10 @@ import lombok.ToString;
 public class User {
 
     @Id
-//    @GeneratedValue(strategy = GenerationType.AUTO)
-    @SequenceGenerator(name="user_generator", sequenceName="user_seq", allocationSize=1)
-    @GeneratedValue(strategy=GenerationType.SEQUENCE, generator="user_generator")
+    @SequenceGenerator(name="user_generator", sequenceName="user_seq", allocationSize= 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "user_generator")
     @Column(name = "id")
-    public Integer id;
+    public int id;
 
     @Column(name = "first_name")
     public String firstName;
@@ -26,7 +25,7 @@ public class User {
     public String lastName;
 
    @Column(name = "telegram_id")
-    public Integer telegramId;
+    public Long telegramId;
 
    @Column(name = "user_name")
     public  String userName;
